requirements.txt
flask
flask-cors
pandas
numpy
sentence-transformers
faiss-cpu
openai
python-dotenv